+++ Instances.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import _ from 'lodash';
+import { stripesShape } from '@folio/stripes-core/src/Stripes'; // eslint-disable-line import/no-unresolved
 
 import SearchAndSort from '@folio/stripes-smart-components/lib/SearchAndSort';
 import { filters2cql, onChangeFilter as commonChangeFilter } from '@folio/stripes-components/lib/FilterGroups';
@@ -307,10 +308,7 @@ class Instances extends React.Component {
 }
 
 Instances.propTypes = {
-  stripes: PropTypes.shape({
-    connect: PropTypes.func.isRequired,
-    locale: PropTypes.string.isRequired,
-  }).isRequired,
+  stripes: stripesShape.isRequired,
   resources: PropTypes.shape({
     records: PropTypes.shape({
       hasLoaded: PropTypes.bool.isRequired,